requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
coloredlogs==15.0.1
cssselect==1.3.0
customtkinter==5.2.2
darkdetect==0.8.0
docopt==0.6.2
et_xmlfile==2.0.0
eutils==0.6.0
filetype==1.2.0
google-ai-generativelanguage==0.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-genai==1.38.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
habanero==2.3.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.76
langchain-google-genai==2.1.12
langchain-text-splitters==0.3.11
langsmith==0.4.28
Levenshtein==0.27.1
lxml==6.0.1
lxml_html_clean==0.4.2
metapub==0.6.4
numpy==2.3.3
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pefile==2023.2.7
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.14.1
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tabulate==0.9.0
tenacity==9.1.2
tkinter-async-execute==1.4.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
websockets==15.0.1
zstandard==0.25.0
